Loner working with Perform later and not Resque
Found a huge bug that Redis was setting the loner names with the redis
namespace for Resque.

The problem was that the names were being set correctly but not being
cleared correctly sometimes.

I was not able to find the rules of when that's happening so I just
moved to `PerformLater.config.resque` without a namespace which will
remove the dependency.

All the specs are passing and I will changed the configuration in the
app after I bump resque